+++ docker_daemon/test_docker_daemon.py
@@ -10,6 +10,7 @@ import mock
 from nose.plugins.attrib import attr
 
 # project
+from checks import AgentCheck
 from tests.checks.common import AgentCheckTest
 from utils.dockerutil import DockerUtil
 
@@ -659,6 +660,43 @@ class TestCheckDockerDaemon(AgentCheckTest):
         ]
         for co in containers:
             self.assertEqual(DockerUtil.container_name_extractor(co[0]), co[1])
+
+    def test_collect_exit_code(self):
+        config = {
+            "init_config": {},
+            "instances": [{
+                "url": "unix://var/run/docker.sock",
+                "collect_exit_codes": True
+            }]
+        }
+        DockerUtil().set_docker_settings(config['init_config'], config['instances'][0])
+
+        expected_service_checks = [
+            (AgentCheck.OK, ['docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest', 'container_name:test-exit-ok']),
+            (AgentCheck.CRITICAL, ['docker_image:nginx:latest', 'image_name:nginx', 'image_tag:latest', 'container_name:test-exit-fail']),
+        ]
+
+        container_ok = self.docker_client.create_container(
+            "nginx:latest", detach=True, name='test-exit-ok', entrypoint='/bin/true')
+        log.debug('start nginx:latest with entrypoint /bin/true')
+        container_fail = self.docker_client.create_container(
+            "nginx:latest", detach=True, name='test-exit-fail', entrypoint='/bin/false')
+        log.debug('start nginx:latest with entrypoint /bin/false')
+        self.docker_client.start(container_ok)
+        self.docker_client.start(container_fail)
+        log.debug('container exited with %s' % self.docker_client.wait(container_ok, 1))
+        log.debug('container exited with %s' % self.docker_client.wait(container_fail, 1))
+        # After the container exits, we need to wait a second so the event isn't too recent
+        # when the check runs, otherwise the event is not picked up
+        from time import sleep
+        sleep(1)
+
+        self.run_check(config)
+        self.docker_client.remove_container(container_ok)
+        self.docker_client.remove_container(container_fail)
+
+        for status, tags in expected_service_checks:
+            self.assertServiceCheck('docker.exit', status=status, tags=tags, count=1)
 
     def test_network_tagging(self):
         expected_metrics = [